add documentation and structure to measure tutorials Signed-off-by: dbulahov <bulahovdenis@gmail.com>
