+++ instance/emf/serializer/pom.xml
@@ -365,14 +365,14 @@
     <dependency>
       <groupId>org.eclipse.daanse</groupId>
       <artifactId>
-        org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.inlinetable</artifactId>
+        org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.measure.inlinetable</artifactId>
       <version>${project.version}</version>
     </dependency>
 
     <dependency>
       <groupId>org.eclipse.daanse</groupId>
       <artifactId>
-        org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.inlinetablewithphysical</artifactId>
+        org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.measure.inlinetablewithphysical</artifactId>
       <version>${project.version}</version>
     </dependency>
 
@@ -393,7 +393,7 @@
     <dependency>
       <groupId>org.eclipse.daanse</groupId>
       <artifactId>
-        org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.measureexpression</artifactId>
+        org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.measure.expression</artifactId>
       <version>${project.version}</version>
     </dependency>
