+++ instance/emf/serializer/test.bndrun
@@ -52,7 +52,7 @@
 	bnd.identity;id='org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.database.column',\
 	bnd.identity;id='org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.database.expressioncolumn',\
 	bnd.identity;id='org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.database.inlinetable',\
-	bnd.identity;id='org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.inlinetablewithphysical',\
+	bnd.identity;id='org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.measure.inlinetablewithphysical',\
 	bnd.identity;id='org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.database.intro',\
 	bnd.identity;id='org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.database.schema',\
 	bnd.identity;id='org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.database.sqlview',\
@@ -92,10 +92,10 @@
 	bnd.identity;id='org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.action.drillthrough',\
 	bnd.identity;id='org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.hierarchywithinnertable',\
 	bnd.identity;id='org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.hierarchywithview',\
-	bnd.identity;id='org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.inlinetable',\
+	bnd.identity;id='org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.measure.inlinetable',\
 	bnd.identity;id='org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.kpi.all',\
 	bnd.identity;id='org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.kpi.virtualcube',\
-	bnd.identity;id='org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.measureexpression',\
+	bnd.identity;id='org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.measure.expression',\
 	bnd.identity;id='org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.namedset.all',\
 	bnd.identity;id='org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.dimension.timedimension',\
 	bnd.identity;id='org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.virtualcube.min',\
@@ -192,15 +192,15 @@
 	org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.function.logic;version='[0.0.1,0.0.2)',\
 	org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.hierarchywithinnertable;version='[0.0.1,0.0.2)',\
 	org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.hierarchywithview;version='[0.0.1,0.0.2)',\
-	org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.inlinetable;version='[0.0.1,0.0.2)',\
-	org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.inlinetablewithphysical;version='[0.0.1,0.0.2)',\
 	org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.kpi.all;version='[0.0.1,0.0.2)',\
 	org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.kpi.intro;version='[0.0.1,0.0.2)',\
 	org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.kpi.parent.ring;version='[0.0.1,0.0.2)',\
 	org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.kpi.virtualcube;version='[0.0.1,0.0.2)',\
 	org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.levelifblankname;version='[0.0.1,0.0.2)',\
 	org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.levelifparentsname;version='[0.0.1,0.0.2)',\
-	org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.measureexpression;version='[0.0.1,0.0.2)',\
+	org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.measure.expression;version='[0.0.1,0.0.2)',\
+	org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.measure.inlinetable;version='[0.0.1,0.0.2)',\
+	org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.measure.inlinetablewithphysical;version='[0.0.1,0.0.2)',\
 	org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.member.identifier;version='[0.0.1,0.0.2)',\
 	org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.namedset.all;version='[0.0.1,0.0.2)',\
 	org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.virtualcube.min;version='[0.0.1,0.0.2)',\
