+++ instance/emf/tutorial/cellformatter/src/main/java/org/eclipse/daanse/rolap/mapping/instance/emf/tutorial/cellformatter/CatalogSupplier.java
@@ -67,6 +67,7 @@ public class CatalogSupplier implements CatalogMappingSupplier {
         query.setTable(table);
 
         CellFormatter cellFormatter = RolapMappingFactory.eINSTANCE.createCellFormatter();
+        cellFormatter.setId("_cellFormatter");
         cellFormatter.setRef("mondrian.rolap.format.CellFormatterImpl");
 
         SumMeasure measure1 = RolapMappingFactory.eINSTANCE.createSumMeasure();
