+++ instance/emf/tutorial/measure/expression/src/main/resources/data/MEASURE_TABLE.csv
