+++ instance/emf/tutorial/measure/inlinetablewithphysical/src/main/resources/data/TOWN.csv
