+++ instance/emf/tutorial/measure/pom.xml
@@ -1,0 +1,31 @@
+<?xml version="1.0"?>
+<!--
+/*********************************************************************
+* Copyright (c) 2025 Contributors to the Eclipse Foundation.
+*
+* This program and the accompanying materials are made
+* available under the terms of the Eclipse Public License 2.0
+* which is available at https://www.eclipse.org/legal/epl-2.0/
+*
+* SPDX-License-Identifier: EPL-2.0
+**********************************************************************/
+-->
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <parent>
+    <groupId>org.eclipse.daanse</groupId>
+    <artifactId>org.eclipse.daanse.rolap.mapping.instance.emf.tutorial</artifactId>
+    <version>${revision}</version>
+  </parent>
+  <packaging>pom</packaging>
+
+  <artifactId>org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.measure</artifactId>
+
+  <modules>
+    <module>expression</module>
+    <module>inlinetable</module>
+    <module>inlinetablewithphysical</module>
+  </modules>
+</project>
